==> Work/Modern Ways/Web Scrapping/Etymology/Product/2-Etymology Language Pair.ipynb <==
### Etymology Language Pair

This notebook is used to etymology data analysis for language pair.

import pandas as pd
import numpy as np
import glob
import nltk
from nltk import word_tokenize
from pathlib import Path
import shutil

#### All Etymology File Concat

#all_etymologeek_file = glob.glob("/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Data/*/*_Etymologeek_Main_And_Additional_Result.xlsx")
#all_etymologeek_file

#df_etymologeek_all = pd.DataFrame()
#for i in all_etymologeek_file:
#    df_var = pd.read_excel(f"{i}")
#    df_etymologeek_all = pd.concat([df_etymologeek_all, df_var], axis=0)
#df_etymologeek_all.drop_duplicates(inplace=True)
#df_etymologeek_all.reset_index(drop=True, inplace=True)
#df_etymologeek_all

#df_etymologeek_all.to_excel("All_Etymology_Language_Pair_Result.xlsx", index=False)

#### Read Etymologeek All Data

df_etymologeek_all = pd.read_excel("/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Data/All/All_Etymology_Language_Pair_Result.xlsx")
df_etymologeek_all

#df_etymologeek_all["language_main"].unique()

#df_etymologeek_all["language"].unique()

lang_1 = "English"  # English, Turkish, German, French, Spanish, Arabic, Russian, Portuguese, Dutch, Italian
lang_2 = "Turkish (tur)"  # English (eng), Turkish (tur), German (deu), French (fra), Spanish (spa), Arabic (ara), Russian (rus), 
                          # Portuguese (por), Dutch (nld), Italian (ita)

Path(f"/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Result/{lang_1.capitalize()}").mkdir(parents=True, exist_ok=True) 

df_lang_pair1 = df_etymologeek_all[(df_etymologeek_all["language_main"] == f"{lang_1.capitalize()}") & (df_etymologeek_all["language"] == f"{lang_2.capitalize()}")]
df_lang_pair1

df_lang_pair1["dict_entry_main"] = df_lang_pair1["dict_entry_main"].apply(lambda x: x.lower())
df_lang_pair1["dict_entry"] = df_lang_pair1["dict_entry"].apply(lambda x: x.lower())
df_lang_pair1 = df_lang_pair1.drop_duplicates(subset="dict_entry_main")
df_lang_pair1.reset_index(drop=True, inplace=True)
df_lang_pair1

#df_lang_pair1.to_excel(f"/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Result/{lang_1.capitalize()}/{lang_1.capitalize()}_And_{lang_2.capitalize()}_Etymologeek_Result1.xlsx", index=False)

#### Read Github Cross Data

##### Part 1

df_github_cross = pd.read_csv("/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Data/All/etymology_cross_check2.csv")
df_github_cross

lang_1 = "Turkish"  # English, Turkish, German, French, Spanish, Arabic, Russian, Portuguese, Dutch, Italian
lang_2 = "Italian"  # English, Turkish, German, French, Spanish, Arabic, Russian, Portuguese, Dutch, Italian

Path(f"/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Data/{lang_1.capitalize()}/{lang_1.capitalize()} {lang_2.capitalize()}").mkdir(parents=True, exist_ok=True)
Path(f"/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Result/{lang_1.capitalize()}/{lang_1.capitalize()} {lang_2.capitalize()}").mkdir(parents=True, exist_ok=True)

def low_func(x):
    try:
        var_out = x.lower()
    except:
        var_out = x
    return var_out 

df_lang1 = df_github_cross[(df_github_cross["lang"] == f"{lang_1.capitalize()}")] 
df_lang1

#df_lang1.to_excel("Turkish_Word_From_Other_Language.xlsx", index=False)

df_lang2 = df_github_cross[(df_github_cross["lang"] == f"{lang_2.capitalize()}")]
df_lang2["term"] = df_lang2["term"].apply(lambda x: x.lower())
#df_lang2.rename(columns={"term":"word"}, inplace=True)
df_lang2

#df_word = pd.read_excel(f"/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/{lang_2.capitalize()}/Result/Word/Merge/Word_Merge2.xlsx")
#df_word = df_word.head(40000)
#df_word = df_word.loc[:,["word"]]
#df_word

#df_lang2 = pd.merge(df_lang2,df_word,how="inner",on="word")
#df_lang2

set_lang1 = set(df_lang1["related_term_id"])
set_lang2= set(df_lang2["related_term_id"])

df_pair_etymology = pd.DataFrame(set_lang1.intersection(set_lang2),columns=["related_term_id"])
df_pair_etymology

df_etymology_merge = pd.merge(df_github_cross,df_pair_etymology,how="inner",on="related_term_id")
df_etymology_merge.drop_duplicates(inplace=True)
df_etymology_merge

df_etymology_merge = df_etymology_merge[(df_etymology_merge["reltype"] == "inherited_from") | (df_etymology_merge["reltype"] == "etymologically_related_to") | (df_etymology_merge["reltype"] == "derived_from") | (df_etymology_merge["reltype"] == "borrowed_from")]
df_etymology_merge

df_etymology_merge = df_etymology_merge[~df_etymology_merge["related_term"].str.contains(fr"(?:\s|^)\*|\-(?:\s|$)", na=True)]
df_etymology_merge

df_lang1_result = df_etymology_merge[(df_etymology_merge["lang"]==f"{lang_1.capitalize()}")]
df_lang1_result.drop_duplicates(inplace=True)
df_lang1_result.reset_index(drop=True, inplace=True)
df_lang1_result

df_lang1_result["term"] = df_lang1_result["term"].apply(lambda x: low_func(x))
df_lang1_result["related_term"] = df_lang1_result["related_term"].apply(lambda x: low_func(x))
df_lang1_result = df_lang1_result.loc[:,["lang","term","related_term_id"]]
df_lang1_result

df_lang2_result = df_etymology_merge[(df_etymology_merge["lang"]==f"{lang_2.capitalize()}")]
df_lang2_result.drop_duplicates(inplace=True)
df_lang2_result.reset_index(drop=True, inplace=True)
df_lang2_result

df_lang2_result["term"] = df_lang2_result["term"].apply(lambda x: low_func(x))
df_lang2_result["related_term"] = df_lang2_result["related_term"].apply(lambda x: low_func(x))
df_lang2_result = df_lang2_result.loc[:,["lang","term","related_term_id"]]
df_lang2_result

df_same_ety = pd.merge(df_lang1_result,df_lang2_result, how="inner", on="related_term_id", suffixes=["_master","_sleeve"])
df_same_ety = df_same_ety[["lang_master","term_master","lang_sleeve","term_sleeve"]]
df_same_ety.drop_duplicates(inplace=True)
df_same_ety

df_same_ety.to_excel(f"/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Result/{lang_1.capitalize()}/{lang_1.capitalize()} {lang_2.capitalize()}/{lang_1.capitalize()}_And_{lang_2.capitalize()}_Etymologeek_Result2.xlsx", index=False)

##### Part 2

df_github_clean = pd.read_csv("/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Data/All/etymology_clean.csv")
df_github_clean

#lang_1 = "English"  # English, Turkish, German, French, Spanish, Arabic, Russian, Portuguese, Dutch, Italian
#lang_2 = "Turkish"  # English, Turkish, German, French, Spanish, Arabic, Russian, Portuguese, Dutch, Italian

df_lang1_lang2 = df_github_clean[(df_github_clean["lang"] == f"{lang_1.capitalize()}") & (df_github_clean["related_lang"] == f"{lang_2.capitalize()}")]
df_lang1_lang2["term"] = df_lang1_lang2["term"].apply(lambda x: low_func(x))
df_lang1_lang2["related_term"] = df_lang1_lang2["related_term"].apply(lambda x: low_func(x))
df_lang1_lang2.reset_index(drop=True, inplace=True)
df_lang1_lang2

#df_word = pd.read_excel(f"/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/{lang_2.capitalize()}/Result/Word/Merge/Word_Merge2.xlsx")
#df_word = df_word.head(40000)
#df_word = df_word.loc[:,["word"]]
#df_word.rename(columns={"word":"related_term"}, inplace=True)
#df_word

#df_lang1_lang2 = pd.merge(df_lang1_lang2,df_word, how="inner", on="related_term")
#df_lang1_lang2

df_lang1_lang2.to_excel(f"/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Result/{lang_1.capitalize()}/{lang_1.capitalize()} {lang_2.capitalize()}/{lang_1.capitalize()}_And_{lang_2.capitalize()}_Etymologeek_Result3.xlsx", index=False)

##### Part 3

df_github_cross = pd.read_csv("/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Data/All/etymology_cross_check2.csv")
df_github_cross

#lang_1 = "Turkish"  # English, Turkish, German, French, Spanish, Arabic, Russian, Portuguese, Dutch, Italian
#lang_2 = "Portuguese"  # English, Turkish, German, French, Spanish, Arabic, Russian, Portuguese, Dutch, Italian

df_lang_pair2 = df_github_cross[(df_github_cross["lang"] == f"{lang_1.capitalize()}") & (df_github_cross["related_lang"] == f"{lang_2.capitalize()}")]
df_lang_pair2.drop_duplicates(inplace=True)
df_lang_pair2.reset_index(drop=True, inplace=True)
df_lang_pair2

df_lang_pair2["term"] = df_lang_pair2["term"].apply(lambda x: low_func(x))
df_lang_pair2["related_term"] = df_lang_pair2["related_term"].apply(lambda x: low_func(x))

#df_word = pd.read_excel(f"/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/{lang_2.capitalize()}/Result/Word/Merge/Word_Merge2.xlsx")
#df_word = df_word.head(40000)
#df_word = df_word.loc[:,["word"]]
#df_word.rename(columns={"word":"related_term"}, inplace=True)
#df_word

#df_lang_pair2 = pd.merge(df_lang_pair2,df_word, how="inner", on="related_term")
#df_lang_pair2

df_lang_pair2 = df_lang_pair2[(df_lang_pair2["reltype"] == "cognate_of") | (df_lang_pair2["reltype"] == "etymologically_related_to") | (df_lang_pair2["reltype"] == "derived_from") | (df_lang_pair2["reltype"] == "borrowed_from")]
df_lang_pair2 = df_lang_pair2.loc[:,["lang","term","reltype","related_lang","related_term"]]
df_lang_pair2.reset_index(drop=True, inplace=True)
df_lang_pair2

df_lang_pair2.to_excel(f"/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Result/{lang_1.capitalize()}/{lang_1.capitalize()} {lang_2.capitalize()}/{lang_1.capitalize()}_And_{lang_2.capitalize()}_Etymologeek_Result4.xlsx", index=False)
