==> Work/Modern Ways/Web Scrapping/Etymology/Product/3-Etymology Cross Check.ipynb <==
### Etymology Cross Check

This notebook is used to cross-check language pairs to get a more common result.

import pandas as pd
import numpy as np
import glob
import re
import nltk
from nltk import word_tokenize

#### Etymologeek Result

lang_1 = "Turkish"  # English, Turkish, German, French, Spanish, Arabic, Russian, Portuguese, Dutch, Italian 
lang_2 = "Italian"  # English, Turkish, German, French, Spanish, Arabic, Russian, Portuguese, Dutch, Italian

if lang_2 == "English":
    lang_3 = "English (eng)"
elif lang_2 == "Turkish":
    lang_3 = "Turkish (tur)"
elif lang_2 == "German":
    lang_3 = "German (deu)"
elif lang_2 == "French":
    lang_3 = "French (fra)"
elif lang_2 == "Spanish":
    lang_3 = "Spanish (spa)"
elif lang_2 == "Arabic":
    lang_3 = "Arabic (ara)"
elif lang_2 == "Russian":
    lang_3 = "Russian (rus)"
elif lang_2 == "Portuguese":
    lang_3 = "Portuguese (por)"
elif lang_2 == "Dutch":
    lang_3 = "Dutch (nld)"
elif lang_2 == "Italian":
    lang_3 = "Italian (ita)"
else:
    pass
#lang_3 = "English (eng)"  # English (eng), Turkish (tur), German (deu), French (fra), Spanish (spa), Arabic (ara), Russian (rus), 
                          # Portuguese (por), Dutch (nld), Italian (ita)
print(f"{lang_1}\n{lang_2}\n{lang_3}")

#Path(f"/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Result/{lang_1.capitalize()}").mkdir(parents=True, exist_ok=True) 

lang_1_file = glob.glob(f"/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Result/*/{lang_1.capitalize()}_And_*_Etymologeek_Result1.xlsx")
lang_1_file

df_lang_1 = pd.DataFrame()
for i in lang_1_file:
    df_var = pd.read_excel(f"{i}")
    df_lang_1 = pd.concat([df_lang_1,df_var], axis=0)
    df_lang_1.reset_index(drop=True, inplace=True)
df_lang_1

lang_2_file = glob.glob(f"/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Result/*/{lang_2.capitalize()}_And_*_Etymologeek_Result1.xlsx")
lang_2_file

df_lang_2 = pd.DataFrame()
for i in lang_2_file:
    df_var2 = pd.read_excel(f"{i}")
    df_lang_2 = pd.concat([df_lang_2,df_var2], axis=0)
    df_lang_2.reset_index(drop=True, inplace=True)
df_lang_2

set_lang_11 = set(df_lang_1["dict_entry"])

set_lang_12 = set(df_lang_2["dict_entry"])

df_cross = pd.DataFrame(set_lang_11.intersection(set_lang_12), columns=["dict_entry"])
df_cross

df_lang_1_cross = pd.merge(df_cross,df_lang_1,how="inner", on="dict_entry")
df_lang_1_cross.drop_duplicates(inplace=True)
df_lang_1_cross = df_lang_1_cross[["language_main","dict_entry_main","dict_entry"]]
df_lang_1_cross.rename(columns={"language_main":"lang_master","dict_entry_main":"term_master"}, inplace=True)
df_lang_1_cross

df_lang_2_cross = pd.merge(df_cross,df_lang_2,how="inner", on="dict_entry")
df_lang_2_cross.drop_duplicates(inplace=True)
df_lang_2_cross = df_lang_2_cross[["language_main","dict_entry_main","dict_entry"]]
df_lang_2_cross.rename(columns={"language_main":"lang_sleeve","dict_entry_main":"term_sleeve"}, inplace=True)
df_lang_2_cross

df_lang_1_lang_2_cross = pd.merge(df_lang_1_cross,df_lang_2_cross, how="inner", on="dict_entry")
df_lang_1_lang_2_cross.drop_duplicates(inplace=True)
df_lang_1_lang_2_cross = df_lang_1_lang_2_cross[["lang_master","term_master","lang_sleeve","term_sleeve"]]
df_lang_1_lang_2_cross

lang_3_file = glob.glob(f"/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Result/*/*_And_{lang_3.capitalize()}_Etymologeek_Result1.xlsx")
lang_3_file

df_lang_3 = pd.DataFrame()
for i in lang_3_file:
    df_var3 = pd.read_excel(f"{i}")
    df_lang_3 = pd.concat([df_lang_3,df_var3], axis=0)
    df_lang_3.reset_index(drop=True, inplace=True)
df_lang_3

set_lang_21 = set(df_lang_1["dict_entry_main"])

set_lang_23 = set(df_lang_3["dict_entry_main"])

df_cross_2 = pd.DataFrame(set_lang_21.intersection(set_lang_23), columns=["dict_entry_main"])
df_cross_2

df_lang_1_lang_3_cross = pd.merge(df_cross_2,df_lang_3,how="inner", on="dict_entry_main")
df_lang_1_lang_3_cross.drop_duplicates(inplace=True)
df_lang_1_lang_3_cross = df_lang_1_lang_3_cross[["language_main","dict_entry_main","language","dict_entry"]]
df_lang_1_lang_3_cross.rename(columns={"language_main":"lang_master","dict_entry_main":"term_master","language":"lang_sleeve","dict_entry":"term_sleeve"}, inplace=True)
df_lang_1_lang_3_cross

df_lang_1_lang_2_lang_3 = pd.concat([df_lang_1_lang_2_cross,df_lang_1_lang_3_cross], axis=0)
df_lang_1_lang_2_lang_3.reset_index(drop=True, inplace=True)
df_lang_1_lang_2_lang_3["term_master"] = df_lang_1_lang_2_lang_3["term_master"].apply(lambda x: x.lower())
df_lang_1_lang_2_lang_3["term_sleeve"] = df_lang_1_lang_2_lang_3["term_sleeve"].apply(lambda x: x.lower())
df_lang_1_lang_2_lang_3

#df_word = pd.read_excel(f"/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/{lang_2.capitalize()}/Result/Word/Merge/Word_Merge2.xlsx")
#df_word = df_word.head(40000)
#df_word = df_word.loc[:,["word"]]
#df_word.rename(columns={"word":"term_sleeve"}, inplace=True)
#df_word

#df_lang_1_lang_2_lang_3 = pd.merge(df_lang_1_lang_2_lang_3,df_word, how="inner", on="term_sleeve")
#df_lang_1_lang_2_lang_3

df_lang_1_lang_2_lang_3.to_excel(f"/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Result/{lang_1.capitalize()}/{lang_1.capitalize()} {lang_2.capitalize()}/{lang_1.capitalize()}_And_{lang_3.capitalize()}_Etymologeek_Result5.xlsx", index=False)
