==> Work/Modern Ways/Web Scrapping/Etymology/Product/4-Language Pair Word Equality.ipynb <==
#### Language Pair Word Equality

This notebook is used to analyze etymology data and word translation data together over the same pronunciation.

#!pip uninstall googletrans
#import googletrans
#from googletrans import Translator

import pandas as pd
import numpy as np
import glob
import re
import nltk
from nltk import word_tokenize
from pathlib import Path
import shutil

lang_1 = "Turkish"  # English, Turkish, German, French, Spanish, Arabic, Russian, Portuguese, Dutch, Italian 
lang_2 = "Italian"  # English, Turkish, German, French, Spanish, Arabic, Russian, Portuguese, Dutch, Italian

if lang_2 == "English":
    lang_3 = "English (eng)"
elif lang_2 == "Turkish":
    lang_3 = "Turkish (tur)"
elif lang_2 == "German":
    lang_3 = "German (deu)"
elif lang_2 == "French":
    lang_3 = "French (fra)"
elif lang_2 == "Spanish":
    lang_3 = "Spanish (spa)"
elif lang_2 == "Arabic":
    lang_3 = "Arabic (ara)"
elif lang_2 == "Russian":
    lang_3 = "Russian (rus)"
elif lang_2 == "Portuguese":
    lang_3 = "Portuguese (por)"
elif lang_2 == "Dutch":
    lang_3 = "Dutch (nld)"
elif lang_2 == "Italian":
    lang_3 = "Italian (ita)"
else:
    pass

print(f"{lang_1}\n{lang_2}\n{lang_3}")

#Path(f"/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Result/{lang_1.capitalize()}").mkdir(parents=True, exist_ok=True) 

def low_func(x):
    try:
        var_out = x.lower()
    except:
        var_out = x
    return var_out 

df_lang11 = pd.read_excel(f"/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Result/{lang_1.capitalize()}/{lang_1.capitalize()} {lang_2.capitalize()}/{lang_1.capitalize()}_And_{lang_3.capitalize()}_Etymologeek_Result5.xlsx")
df_lang11

df_lang12 = pd.read_excel(f"/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Result/{lang_1.capitalize()}/{lang_1.capitalize()} {lang_2.capitalize()}/{lang_1.capitalize()}_And_{lang_2.capitalize()}_Etymologeek_Result2.xlsx")
df_lang12

df_lang13 = pd.read_excel(f"/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Result/{lang_1.capitalize()}/{lang_1.capitalize()} {lang_2.capitalize()}/{lang_1.capitalize()}_And_{lang_2.capitalize()}_Etymologeek_Result3.xlsx")
df_lang13 = df_lang13[["lang","term","related_lang","related_term"]]
df_lang13.rename(columns={"lang":"lang_master","term":"term_master","related_lang":"lang_sleeve","related_term":"term_sleeve"}, inplace=True)
df_lang13["term_master"] = df_lang13["term_master"].apply(lambda x: low_func(x))
df_lang13["term_sleeve"] = df_lang13["term_sleeve"].apply(lambda x: low_func(x))
df_lang13

df_lang14 = pd.read_excel(f"/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Result/{lang_1.capitalize()}/{lang_1.capitalize()} {lang_2.capitalize()}/{lang_1.capitalize()}_And_{lang_2.capitalize()}_Etymologeek_Result4.xlsx")
df_lang14 = df_lang14[["lang","term","related_lang","related_term"]]
df_lang14.rename(columns={"lang":"lang_master","term":"term_master","related_lang":"lang_sleeve","related_term":"term_sleeve"}, inplace=True)
df_lang14["term_master"] = df_lang14["term_master"].apply(lambda x: low_func(x))
df_lang14["term_sleeve"] = df_lang14["term_sleeve"].apply(lambda x: low_func(x))
df_lang14

df_ety_concat = pd.concat([df_lang11,df_lang12,df_lang13,df_lang14], axis=0)
df_ety_concat.drop_duplicates(inplace=True)
df_ety_concat.reset_index(drop=True, inplace=True)
df_ety_concat

df_ety_concat.to_excel(f"/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Data/{lang_1.capitalize()}/{lang_1.capitalize()} {lang_2.capitalize()}/{lang_1.capitalize()}_All_Word_From_{lang_2.capitalize()}_Etymology.xlsx", index=False)

df_master_unique = pd.DataFrame()
df_master_unique["term_master"] = df_ety_concat["term_master"].unique()
df_master_unique.rename(columns={"term_master":"term"}, inplace=True)
df_master_unique

df_master_unique.to_excel(f"/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Data/{lang_1.capitalize()}/{lang_1.capitalize()} {lang_2.capitalize()}/{lang_1.capitalize()}_Unique_Word_From_{lang_2.capitalize()}_Etymology.xlsx", index=False)

#df_word = pd.read_excel(f"/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/{lang_1.capitalize()}/Result/Word/Merge/Word_Merge2.xlsx")
#df_word = df_word.head(40000)
##df_word = df_word.loc[:,["word"]]
#df_word.rename(columns={"word":"term"}, inplace=True)
#df_word

#df_master_unique_word_merge = pd.merge(df_word,df_master_unique, how="inner", on="term")
#df_master_unique_word_merge.sort_values(by="frequency", inplace=True, ascending=False)
#df_master_unique_word_merge.reset_index(drop=True, inplace=True)
#df_master_unique_word_merge

#df_master_unique_word_merge = df_master_unique_word_merge.loc[:,["term"]]
#df_master_unique_word_merge.rename(columns={"term":f"{lang_1.lower()}_word"}, inplace=True)
#df_master_unique_word_merge

#df_master_unique_word_merge.to_excel(f"/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Data/{lang_1.capitalize()}/{lang_1.capitalize()}_Unique_Word_From_{lang_2.capitalize()}_Etymology.xlsx", index=False)

#### Read Translate File 

# this data from Google Sheets tranlate result
df_pair_translate = pd.read_excel(f"/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Data/{lang_1.capitalize()}/{lang_1.capitalize()} {lang_2.capitalize()}/{lang_1.capitalize()} {lang_2.capitalize()} Language Pair Translate.xlsx")
df_pair_translate

df_pair_translate[f"{lang_2.lower()}_word"] = df_pair_translate[f"{lang_2.lower()}_word"].apply(lambda x: low_func(x))
df_pair_translate

df_sleeve_unique = pd.DataFrame()
df_sleeve_unique["term_sleeve"] = df_ety_concat["term_sleeve"].unique()
df_sleeve_unique.rename(columns={"term_sleeve":f"{lang_2.lower()}_word"}, inplace=True)
df_sleeve_unique[f"{lang_2.lower()}_word"] = df_sleeve_unique[f"{lang_2.lower()}_word"].apply(lambda x: low_func(x))
df_sleeve_unique.drop_duplicates(inplace=True)
df_sleeve_unique.reset_index(drop=True, inplace=True)
df_sleeve_unique

df_translate_sleeve_merge = pd.merge(df_pair_translate,df_sleeve_unique,how="inner",on=f"{lang_2.lower()}_word")
df_translate_sleeve_merge.rename(columns={f"{lang_1.lower()}_word":"dict_entry_main"}, inplace=True)
df_translate_sleeve_merge

df_translate_sleeve_merge.to_excel(f"/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Result/{lang_1.capitalize()}/{lang_1.capitalize()} {lang_2.capitalize()}/{lang_1.capitalize()}_And_{lang_2.capitalize()}_Etymologeek_Concat_Result1.xlsx", index=False)

#df_translate_sleeve_merge.to_excel(f"{lang_1.capitalize()}_And_{lang_3.capitalize()}_Etymologeek_Result_All.xlsx", index=False)
