==> Work/Modern Ways/Web Scrapping/Etymology/Product/5-Etymology TVS File Analysis.ipynb <==
### Etymology TVS File

This notebook is used to analyze etimology TVS data.

import pandas as pd
import numpy as np
import glob
import re
import nltk
from nltk import word_tokenize
from pathlib import Path
import shutil

#### Read TVS And Convert CSV

#df_tvs = pd.read_csv("/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Data/All/etymwn.tsv", sep="\t")
#df_tvs.rename(columns={"aaq: Pawanobskewi":"column_1","rel:etymological_origin_of":"column_2","eng: Penobscot":"column_3"}, inplace=True)
#df_tvs

#def low_func(x):
#    try:
#        var_out = x.lower()
#    except:
#        var_out = x
#    return var_out 

#def strip_func(x):
#    try:
#        var_out = x.strip()
#    except:
#        var_out = x
#    return var_out 

#df_col_1 = df_tvs["column_1"].str.split(":", n=-1, expand=True)
#df_col_1 = df_col_1.iloc[:,[0,1]]
#df_col_1.rename(columns={0:"lang",1:"term"}, inplace=True)
#df_col_2 = df_tvs["column_2"].str.split(":", n=-1, expand=True)
#df_col_2 = df_col_2.iloc[:,[1]]
#df_col_2.rename(columns={1:"reltype"}, inplace=True)
#df_col_3 = df_tvs["column_3"].str.split(":", n=-1, expand=True)
#df_col_3 = df_col_3.iloc[:,[0,1]]
#df_col_3.rename(columns={0:"related_lang",1:"related_term"}, inplace=True)

#df_tvs_expand =  pd.concat([df_col_1,df_col_2,df_col_3], axis=1)
#df_tvs_expand

#df_tvs_expand["lang"] = df_tvs_expand["lang"].apply(lambda x: low_func(x))
#df_tvs_expand["lang"] = df_tvs_expand["lang"].apply(lambda x: strip_func(x))
#df_tvs_expand["term"] = df_tvs_expand["term"].apply(lambda x: low_func(x))
#df_tvs_expand["term"] = df_tvs_expand["term"].apply(lambda x: strip_func(x))
#df_tvs_expand["reltype"] = df_tvs_expand["reltype"].apply(lambda x: low_func(x))
#df_tvs_expand["reltype"] = df_tvs_expand["reltype"].apply(lambda x: strip_func(x))
#df_tvs_expand["related_lang"] = df_tvs_expand["related_lang"].apply(lambda x: low_func(x))
#df_tvs_expand["related_lang"] = df_tvs_expand["related_lang"].apply(lambda x: strip_func(x))
#df_tvs_expand["related_term"] = df_tvs_expand["related_term"].apply(lambda x: low_func(x))
#df_tvs_expand["related_term"] = df_tvs_expand["related_term"].apply(lambda x: strip_func(x))

#df_tvs_expand.to_csv("/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Data/All/etymwn.csv", index=False)

#### Read CSV File

##### Part 1

df_etymology = pd.read_csv("/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Data/All/etymwn.csv")
df_etymology

lang_1 = "tur"  # tur, eng, fra, deu, por, spa, ita, nld, ara, rus
lang_2 = "ita"  # tur, eng, fra, deu, por, spa, ita, nld, ara, rus

if lang_1 == "tur":
    lang_3 = "Turkish"
elif lang_1 == "eng":
    lang_3 = "English"
elif lang_1 == "fra":
    lang_3 = "French"
elif lang_1 == "deu":
    lang_3 = "German"
elif lang_1 == "por":
    lang_3 = "Portuguese"
elif lang_1 == "spa":
    lang_3 = "Spanish"
elif lang_1 == "ita":
    lang_3 = "Italian"
elif lang_1 == "nld":
    lang_3 = "Dutch"
elif lang_1 == "ara":
    lang_3 = "Arabic"
elif lang_1 == "rus":
    lang_3 = "Russian"
else:
    pass

if lang_2 == "tur":
    lang_4 = "Turkish"
elif lang_2 == "eng":
    lang_4 = "English"
elif lang_2 == "fra":
    lang_4 = "French"
elif lang_2 == "deu":
    lang_4 = "German"
elif lang_2 == "por":
    lang_4 = "Portuguese"
elif lang_2 == "spa":
    lang_4 = "Spanish"
elif lang_2 == "ita":
    lang_4 = "Italian"
elif lang_2 == "nld":
    lang_4 = "Dutch"
elif lang_2 == "ara":
    lang_4 = "Arabic"
elif lang_2 == "rus":
    lang_4 = "Russian"
else:
    pass

print(f"{lang_1}\n{lang_2}\n{lang_3}\n{lang_4}")

#Path(f"/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Result/{lang_3.capitalize()}").mkdir(parents=True, exist_ok=True) 

df_lang1 = df_etymology[df_etymology.lang == f"{lang_1.lower()}"]
df_lang1.reset_index(drop=True, inplace=True)
df_lang1 = df_lang1[~(df_lang1["related_lang"] == f"{lang_1.lower()}")]
df_lang1 = df_lang1[~(df_lang1["related_lang"] == f"{lang_2.lower()}")]
df_lang1

df_lang2 = df_etymology[df_etymology.lang == f"{lang_2.lower()}"]
df_lang2.reset_index(drop=True, inplace=True)
df_lang2 = df_lang2[~(df_lang2["related_lang"] == f"{lang_1.lower()}")]
df_lang2 = df_lang2[~(df_lang2["related_lang"] == f"{lang_2.lower()}")]
df_lang2

df_lang1_lang2 = pd.merge(df_lang1,df_lang2,how="inner",on=["related_lang","related_term"])
df_lang1_lang2.drop_duplicates(inplace=True)
df_lang1_lang2.reset_index(drop=True, inplace=True)
df_lang1_lang2

df_lang1_lang2_result1 = df_lang1_lang2.loc[:,["term_x","term_y"]]
df_lang1_lang2_result1.rename(columns={"term_x":"dict_entry_main","term_y":f"{lang_4.lower()}_word"}, inplace=True)
df_lang1_lang2_result1.reset_index(drop=True, inplace=True)
df_lang1_lang2_result1

##### Part 2

df_lang1_lang2_result2 = df_etymology[(df_etymology["lang"] == f"{lang_1.lower()}") & (df_etymology["related_lang"] == f"{lang_2.lower()}")]
df_lang1_lang2_result2

df_lang1_lang2_result2 = df_lang1_lang2_result2.loc[:,["term","related_term"]]
df_lang1_lang2_result2.rename(columns={"term":"dict_entry_main","related_term":f"{lang_4.lower()}_word"}, inplace=True)
df_lang1_lang2_result2.reset_index(drop=True, inplace=True)
df_lang1_lang2_result2

##### Concat Part 1 And Part 2

#df_part1_part2_concat = pd.concat([df_lang1_lang2_word_check,df_lang1_lang2_word_check2], axis=0)
df_part1_part2_concat = pd.concat([df_lang1_lang2_result1,df_lang1_lang2_result2], axis=0)
df_part1_part2_concat.drop_duplicates(inplace=True)
df_part1_part2_concat.reset_index(drop=True, inplace=True)
df_part1_part2_concat

df_part1_part2_concat.to_excel(f"/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Result/{lang_3.capitalize()}/{lang_3.capitalize()} {lang_4.capitalize()}/{lang_3.capitalize()}_And_{lang_4.capitalize()}_Etymologeek_Concat_Result2.xlsx", index=False)
