==> Work/Modern Ways/Web Scrapping/Etymology/Product/6-Etymology Concat Result Merge.ipynb <==
### Etymology Concat Result Merge

This notebook is used to combine all etymology file result into one file.

#!pip install --upgrade pip

#!pip install regex

import pandas as pd
import numpy as np
import glob
import re
import nltk
from nltk import word_tokenize
from pathlib import Path
import shutil

lang_1 = "Turkish"  # English, Turkish, German, French, Spanish, Arabic, Russian, Portuguese, Dutch, Italian 
lang_2 = "Italian"  # English, Turkish, German, French, Spanish, Arabic, Russian, Portuguese, Dutch, Italian

if lang_2 == "English":
    lang_3 = "English (eng)"
elif lang_2 == "Turkish":
    lang_3 = "Turkish (tur)"
elif lang_2 == "German":
    lang_3 = "German (deu)"
elif lang_2 == "French":
    lang_3 = "French (fra)"
elif lang_2 == "Spanish":
    lang_3 = "Spanish (spa)"
elif lang_2 == "Arabic":
    lang_3 = "Arabic (ara)"
elif lang_2 == "Russian":
    lang_3 = "Russian (rus)"
elif lang_2 == "Portuguese":
    lang_3 = "Portuguese (por)"
elif lang_2 == "Dutch":
    lang_3 = "Dutch (nld)"
elif lang_2 == "Italian":
    lang_3 = "Italian (ita)"
else:
    pass

print(f"{lang_1}\n{lang_2}\n{lang_3}")

# custom character alphabet for word
tr = re.compile(r"[abcçdefgğhıijklmnoöprsştuüvyzqxw]+", re.IGNORECASE|re.UNICODE) # Turkish filters='!"#$%&()*+,-./:;<=>?@[\\]^_`{|}~\t\n0123456789'
en = re.compile(r"[abcdefghıijklmnopqrstxuvwyz]+", re.IGNORECASE|re.UNICODE) # English
nl = re.compile(r"[abcdefghıijklmnopqrstxuvwyzāăēĕīĭōŏūŭ]+", re.IGNORECASE|re.UNICODE) # Dutch (Flemenk)
fr = re.compile(r"[abcçdefghıijklmnopqrstxuvwyzàâæèéêëîïôœùûüÿ]+", re.IGNORECASE|re.UNICODE) # French
de = re.compile(r"[abcdefghıijklmnopqrstxuvwyzäöüß]+", re.IGNORECASE|re.UNICODE) # German
es = re.compile(r"[abcdefghıijklmnopqrstxuvwyzñáéíóú]+", re.IGNORECASE|re.UNICODE) # Spanish (¿¡)
pt = re.compile(r"[abcçdefghıijklmnopqrstxuvwyzàáâãéêíóôõú]+", re.IGNORECASE|re.UNICODE) # Portuguese
it = re.compile(r"[abcdefghıijklmnopqrstxuvwyzàéèìòùî]+", re.IGNORECASE|re.UNICODE) # Italian
ar = re.compile(r"[ٿصؼۤڳڲؿڎػڠجڿ٬ٸؽؒؓطۄڀۂؘؔتٚڛےٝڜؖڦ٫ډ۰زۇٖۀ،لۓعٮێڔ۶ؚۧۜڤۏإٞٷؗۖ؈ژۣؕؑٴأۻڸۺگاڴڹۯ؉ْڌ؍ي؟ـٟړۅؐڶُىڽېًۢؠضۚڄٛڏٱۦ٩س٦ڼڂٔۘ٠ښٌٍ۬ٳ۾ٲږذۋٵٜ٘ڞڅںٗهڣۿپڒۥۗڋیؙم؞ثۨٹڵڪظٶۭ١ڭەڨحٕ؎ٺڷٰ۪۫ڻڥۛڑڟټآڡغګ؊ّٯڧڮ؏ۮ؋ؤ٪ؾڗۼق۟دکوِڰڐۃ۽ہفرڇچڝ۴بۈٽڕۡھةٓڃئ؛ڬٙڙڢڱۊَۆۉځ۠ۍۑۙڊنءڈٻشڍ؇۵كخ\ا]+", re.IGNORECASE|re.UNICODE) # Arabic

# English, Turkish, German, French, Spanish, Arabic, Portuguese, Dutch, Italian 
if lang_1 == "English":
    custom_alp = en
elif lang_1 == "Turkish":
    custom_alp = tr
elif lang_1 == "German":
    custom_alp = de
elif lang_1 == "French":
    custom_alp = fr
elif lang_1 == "Spanish":
    custom_alp = es
elif lang_1 == "Arabic":
    custom_alp = ar
elif lang_1 == "Portuguese":
    custom_alp = pt
elif lang_1 == "Dutch":
    custom_alp = nl
elif lang_1 == "Italian":
    custom_alp = it
else:
    pass
#custom_alp = tr  # en, tr, de, fr, es, ar, pt, nl, it
print(f"{lang_1}\n{custom_alp}") 

def clean_text(text):
    #text_clean = re.findall(custom_alp, text)
    text_clean = re.findall(custom_alp, str(text))
    text_result = "".join(text_clean)
    return text_result

def low_strip_func(x):
    try:
        var_low = str(x).lower()
        var_out = var_low .strip()
    except:
        var_out = str(x)
    return var_out 

# Language Pair Word Equality
df_concat1 = pd.read_excel(f"/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Result/{lang_1.capitalize()}/{lang_1.capitalize()} {lang_2.capitalize()}/{lang_1.capitalize()}_And_{lang_2.capitalize()}_Etymologeek_Concat_Result1.xlsx")
df_concat1

# Etymology TVS File Analysis
df_concat2 = pd.read_excel(f"/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Result/{lang_1.capitalize()}/{lang_1.capitalize()} {lang_2.capitalize()}/{lang_1.capitalize()}_And_{lang_2.capitalize()}_Etymologeek_Concat_Result2.xlsx")
df_concat2

# Etymolojitürk Nişanyan Concat Result from excel file process
df_concat3 = pd.read_excel(f"/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Result/{lang_1.capitalize()}/{lang_1.capitalize()} {lang_2.capitalize()}/{lang_1.capitalize()}_And_{lang_2.capitalize()}_Etymolojitürk_Nisanyan_Concat_Result3.xlsx")
df_concat3

df_merge_result = pd.concat([df_concat1,df_concat2,df_concat3], axis=0)
df_merge_result["dict_entry_main"] = df_merge_result["dict_entry_main"].apply(lambda x: low_strip_func(x))
df_merge_result[f"{lang_2.lower()}_word"] = df_merge_result[f"{lang_2.lower()}_word"].apply(lambda x: low_strip_func(x))
df_merge_result.drop_duplicates(inplace=True)
df_merge_result.reset_index(drop=True, inplace=True)
df_merge_result

df_merge_result.loc[:,"dict_entry_main"] = df_merge_result.loc[:,"dict_entry_main"].apply(lambda x: clean_text(x))
df_merge_result.dropna(inplace=True)
df_merge_result.drop(df_merge_result[df_merge_result.loc[:,"dict_entry_main"] == ""].index, inplace=True)
df_merge_result.reset_index(drop=True, inplace=True)
df_merge_result

df_merge_result["dict_entry_main"].nunique()

# master data
df_merge_result.to_excel(f"/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Result/{lang_1.capitalize()}/{lang_1.capitalize()} {lang_2.capitalize()}/{lang_1.capitalize()}_And_{lang_3.capitalize()}_Etymologeek_Result_All.xlsx", index=False)

### Other Workspace

#### Compare All Cross Word With Result

df_all_cross = pd.read_excel(f"/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Data/{lang_1.capitalize()}/{lang_1.capitalize()} {lang_2.capitalize()}/{lang_1.capitalize()}_All_Word_From_{lang_2.capitalize()}_Etymology.xlsx")
df_all_cross["term_master"] = df_all_cross["term_master"].apply(lambda x: low_strip_func(x))
df_all_cross["term_sleeve"] = df_all_cross["term_sleeve"].apply(lambda x: low_strip_func(x))
df_all_cross

df_all_cross.rename(columns={"term_master":"dict_entry_main", "term_sleeve":f"{lang_2.lower()}_word"}, inplace=True)
df_all_cross

df_merge_result

df_collective = pd.merge(df_all_cross,df_merge_result, how="inner", on="dict_entry_main")
df_collective = df_collective.loc[:,["lang_master","dict_entry_main","lang_sleeve",f"{lang_2.lower()}_word_x"]]
df_collective.rename(columns={f"{lang_2.lower()}_word_x":f"{lang_2.lower()}_word"}, inplace=True)
df_collective

df_not_collective = pd.concat([df_all_cross,df_collective]).drop_duplicates(keep=False)
df_not_collective.reset_index(drop=True, inplace=True)
df_not_collective

df_not_collective.to_excel(f"/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Result/{lang_1.capitalize()}/\
{lang_1.capitalize()} {lang_2.capitalize()}/{lang_1.capitalize()}_And_{lang_2.capitalize()}_Etymologeek_Additional_Result.xlsx", index=False)

#### English French Unique Turkish Words

df_english = pd.read_excel("/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Result/Turkish/Turkish English/Turkish_English_Shared_Vocabulary.xlsx")
df_english

df_french = pd.read_excel("/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Result/Turkish/Turkish French/Turkish_French_Shared_Vocabulary.xlsx")
df_french

set_english = set(df_english.dict_entry_main)
set_french = set(df_french.dict_entry_main)

df_english_french = pd.DataFrame(set_english.union(set_french), columns=["dict_entry_main"])
df_english_french

df_etymology_all = pd.read_excel(f"/media/kurubal/SSD/Data Scientist/Work/Modern Ways/Project/Web Scrapping/Result/{lang_1.capitalize()}/{lang_1.capitalize()} {lang_2.capitalize()}/{lang_1.capitalize()}_And_{lang_3.capitalize()}_Etymologeek_Result_All.xlsx")
df_etymology_all

df_merge_result = pd.merge(df_etymology_all,df_english_french, how="inner", on="dict_entry_main")
df_merge_result.drop_duplicates(inplace=True)
df_merge_result

df_merge_result.to_excel("Turkish_Italian_Result.xlsx", index=False)
